Run "bundle install" by ruby/setup-ruby
`bundle install --without development` doesn't work because `--without`
option was removed. Instead, we need to run `bundle config set without
'development'`.

However, in the first place, we don't need the option because we don't
have 'development' group in out Gemfiles.

So, we just need to run `bundle install`, and ruby/setup-ruby does it.